cdk/pyproject.toml
[tool.poetry]
name = "cloud-midterm-cdk"
version = "0.1.0"
description = ""
authors = ["Chayapol Parisonyu <cpari@sertiscorp.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
